pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "pandas>=2.3.1",
    "openpyxl>=3.1.5",
    "xlsxwriter>=3.2.5",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "google-genai>=1.27.0",
    "pillow>=11.3.0",
    "sift-stack-py>=0.7.0",
    "reportlab>=4.4.3",
    "matplotlib>=3.10.3",
    "requests>=2.32.4",
    "twilio>=9.7.0",
    "scipy>=1.16.1",
    "numpy>=2.3.2",
]
